+++ src/components/ProgressBar/ProgressBar.styled.tsx
@@ -2,6 +2,7 @@ import styled from "styled-components";
 
 import { ProgressBarMetaProps, ProgressBarProps } from "./ProgressBar";
 import { convertToFictoanColor } from "../../utils/helpers";
+import { ColourPropTypes } from "../Element/constants";
 
 
 export const ProgressBarMetaStyled = styled.div`
@@ -40,21 +41,21 @@ export const ProgressBarStyled = styled.progress`
         position   : relative;
         flex       : 1 1 100%;
         height     : ${(props: ProgressBarProps) => props.height};
-        background : ${(props: ProgressBarProps) => convertToFictoanColor(props.barBg) || props.theme.progressBar.bg};
+        background : ${(props: ProgressBarProps) => convertToFictoanColor(props.barBg as ColourPropTypes) || props.theme.progressBar.bg};
         box-shadow : 0 2px 8px -2px hsla(0, 0, 0, 0.24) inset;
     }
 
     &[value]::-webkit-progress-bar {
-        background : ${(props: ProgressBarProps) => convertToFictoanColor(props.barBg) || props.theme.progressBar.bg};
+        background : ${(props: ProgressBarProps) => convertToFictoanColor(props.barBg as ColourPropTypes) || props.theme.progressBar.bg};
         box-shadow : 0 2px 8px -2px hsla(0, 0, 0, 0.24) inset;
     }
 
     &::-webkit-progress-value {
-        background         : ${(props: ProgressBarProps) => convertToFictoanColor(props.barFill) || props.theme.progressBar.fill};
+        background         : ${(props: ProgressBarProps) => convertToFictoanColor(props.barFill as ColourPropTypes) || props.theme.progressBar.fill};
     }
 
     &[value]::-moz-progress-bar {
-        background    : ${(props: ProgressBarProps) => convertToFictoanColor(props.barFill) || props.theme.progressBar.fill};
+        background    : ${(props: ProgressBarProps) => convertToFictoanColor(props.barFill as ColourPropTypes) || props.theme.progressBar.fill};
     }
 
     &.shape-rounded::-webkit-progress-value, &.shape-rounded, &.shape-rounded::-webkit-progress-bar {